Moved articles into components
