+++ app/projects/page.tsx
@@ -1,4 +1,3 @@
-// import ProjectsContent from '@/app/projects/projects_content'
 import ProjectsContent from '@/components/articles/projects'
 import { Metadata } from 'next'
