+++ components/articles/about/index.tsx
@@ -1,14 +1,13 @@
+'use server'
 import { MDXRemote } from 'next-mdx-remote/rsc'
 import {
   HRGradient,
-  TextLargeBoldCenter,
   BackToHome,
   BackToTop,
-  Ribbon
 } from '@/components/layouts'
 import styles from '@/app/layout.module.css'
 
-export const dynamic = 'force-dynamic'
+// export const dynamic = 'force-dynamic'
 
 export default async function AboutContent() {
