+++ components/articles/contact/index.tsx
@@ -1,14 +1,15 @@
+'use server'
 import {
   HRGradient,
   BackToHome,
   BackToTop
 } from '@/components/layouts'
-import RepoCards from '@/components/cards/RepoCards/RepoCards'
+import { ContactForm } from '@/components/elements'
 import styles from '@/app/layout.module.css'
 
-export const dynamic = 'force-dynamic'
+// export const dynamic = 'force-dynamic'
 
-export default async function ProjectsContent() {
+export default async function ContactContent() {
 
   return (
     <div className={styles.container}>
@@ -23,15 +24,15 @@ export default async function ProjectsContent() {
             </div>
           </div> */}
 
-          <HRGradient />
-
           <h2 className='text-center'>
-            Read about some of my most popular projects:
+            Hi! I&#39;m Nathan, a.k.a StoneyDSP.
           </h2>
 
           <HRGradient />
 
-          <RepoCards />
+          <ContactForm />
+
+          <HRGradient />
 
           <HRGradient />
