+++ components/articles/privacy-policy/index.tsx
@@ -1,3 +1,4 @@
+'use server'
 import {
   HRGradient,
   BackToHome,
@@ -5,7 +6,7 @@ import {
 } from '@/components/layouts'
 import styles from '@/app/layout.module.css'
 
-export const dynamic = 'force-dynamic'
+// export const dynamic = 'force-dynamic'
 
 export default async function PrivacyPolicyContent() {
