+++ components/articles/terms-of-service/index.tsx
@@ -1,3 +1,4 @@
+'use server'
 import {
   HRGradient,
   TextLargeBoldCenter,
@@ -6,7 +7,7 @@ import {
 } from '@/components/layouts'
 import styles from '@/app/layout.module.css'
 
-export const dynamic = 'force-dynamic'
+// export const dynamic = 'force-dynamic'
 
 export default async function TermsOfServiceContent() {
